tools/quality-guardian/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quality-guardian"
version = "1.0.0"
description = "Ensure high-quality academic writing with automated checks"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Deep Biblio Tools Team", email = "team@deepbiblio.tools"}
]
keywords = ["quality", "writing", "academic", "grammar", "style", "checker"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "language-tool-python>=2.7",
    "textstat>=0.7",
    "nltk>=3.8",
    "spacy>=3.5",
    "rich>=13.0",
    "PyYAML>=6.0",
    "beautifulsoup4>=4.12",
    "markdown>=3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
quality-guardian = "quality_guardian.cli:main"

[project.urls]
Homepage = "https://github.com/deep-biblio-tools/deep-biblio-tools"
Documentation = "https://deep-biblio-tools.readthedocs.io/tools/quality-guardian"
"Bug Tracker" = "https://github.com/deep-biblio-tools/deep-biblio-tools/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/quality_guardian"]
